Decode: let the symmetric-memory all-reduce run inside CUDA graphs, and sort the MoE routing ids as int32
The graph guard was too wide. A rendezvous is a collective and cannot run during a capture, but replaying a
kernel against a buffer that was rendezvoused before the capture is fine: measured in isolation at tp2, a 2 MiB
all-reduce takes 38.6 us on NCCL and 22.3 us on the two-shot symmetric-memory kernel, both inside a graph. Only
the call that has to allocate now falls back to NCCL. End to end with cuda graphs and compile level 1, Qwen3-8B
tp2 at 256 concurrent: 14,001 -> 14,991 tok/s (+7%); Qwen3-30B-A3B tp4 at 128: 6,616 -> 6,777 (+2%).

The router's expert ids fit in int32 and a 32-bit radix sort makes half the passes of a 64-bit one: 29 us
against 41 us per layer at decode shapes, where that sort is ~8% of the step. Outputs and gradients are
bit-identical (check_moe_sort.py: 0.0 max abs diff on the output, on d_hidden and on d_gate_up_proj).

+++ src/transformers/integrations/moe.py
@@ -387,7 +387,9 @@ def grouped_mm_experts_forward(
 
     # S is the number of selected tokens-experts pairs (S = num_tokens * num_top_k)
     sample_weights = top_k_weights.reshape(-1)  # (S,)
-    expert_ids = top_k_index.reshape(-1)  # (S,)
+    # int32 keys: the router's ids are far below 2**31 and a 32-bit radix sort makes half the passes of a 64-bit one,
+    # which at decode shapes (1024 pairs) is 29 us against 41 us per layer, the sort being ~8% of a MoE decode step
+    expert_ids = top_k_index.reshape(-1).to(torch.int32)  # (S,)
 
     # Sort by expert for grouped processing
     expert_ids_g, perm = torch.sort(expert_ids)
